% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRUM_Rscripts.R
\name{pi_reference}
\alias{pi_reference}
\title{Mapping of Entries of Pi to Latent Attribute Classes}
\usage{
pi_reference(K)
}
\arguments{
\item{K}{A \code{numeric}, denoting the number of attributes.}
}
\value{
A \code{matrix}, where row c represents the latent class
corresponding to entry c of pi
}
\description{
Obtains mapping of entries of pi to latent attribute classes.
}
\examples{
\dontrun{
###################################
### Generate random data from rRUM
### and recover parameters with
### Gibbs sampler
###################################
set.seed(217)
# Define Simulation Parameters
N <- 1000 # number of individuals
J <- 6 # number of items
K <- 2 # number of attributes
As <- pi_reference(K) # matrix where rows represent attribute classes
pis <- c(.1, .2, .3, .4) # pi
Q <- rbind(c(1,0),
           c(0,1),
           c(1,0),
           c(0,1),
           c(1,1),
           c(1,1))
pistar <- rep(.9, J) # pistar
rstar <- .5 * Q # rstar
alpha <- As[sample(1:(K ^ 2), N, replace = TRUE, pis),]
# Simulate data
Y <- simcdm::sim_rrum(Q, rstar, pistar, alpha)
# Recover simulation parameters with Gibbs Sampler
Gibbs.out <- rrum(Y, Q)
burnin <- 1:5000 # iterations to be discarded from chain as burnin
rstar.mean <- with(Gibbs.out, apply(RSTAR[,,-burnin], c(1, 2), mean))
pistar.mean <- with(Gibbs.out, apply(PISTAR[,-burnin], 1, mean))
pis.mean <- with(Gibbs.out, apply(PI[,-burnin], 1 ,mean))
}
}
\references{
Culpepper, S. A. & Hudson, A. (In Press). An improved strategy for Bayesian 
estimation of the reduced reparameterized unified model. Applied
Psychological Measurement.

Hudson, A., Culpepper, S. A., & Douglas, J. (2016, July). Bayesian estimation
of the generalized NIDA model with Gibbs sampling. Paper presented at the
annual International Meeting of the Psychometric Society, Asheville, North
Carolina.
}
\author{
Steven Andrew Culpepper, Aaron Hudson
}
