% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrum-package.R
\docType{package}
\name{rrum-package}
\alias{rrum-package}
\title{rrum: Bayesian Estimation of the Reduced Reparameterized Unified Model with Gibbs Sampling}
\description{
Implementation of Gibbs sampling algorithm for Bayesian Estimation of the Reduced Reparameterized Unified Model ('rrum'), described by Culpepper and Hudson (2017) \doi{ 10.1177/0146621617707511}.
}
\details{
Implemention of a Bayesian estimation for the
reduced Reparameterized Unified Model (rRUM).
}
\references{
Culpepper, S. A. & Hudson, A. (In Press). An improved strategy for Bayesian
estimation of the reduced reparameterized unified model. Applied
Psychological Measurement.

Hudson, A., Culpepper, S. A., & Douglas, J. (2016, July). Bayesian estimation
of the generalized NIDA model with Gibbs sampling. Paper presented at the
annual International Meeting of the Psychometric Society, Asheville, North
Carolina.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/rrum/}
  \item \url{https://github.com/tmsalab/rrum}
  \item Report bugs at \url{https://github.com/tmsalab/rrum/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
  \item Aaron Hudson \email{awhudson@uw.edu} (\href{https://orcid.org/0000-0002-9731-2224}{ORCID}) [copyright holder]
}

}
\keyword{internal}
