% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrum-package.R
\docType{package}
\name{rrum-package}
\alias{rrum-package}
\alias{_PACKAGE}
\title{rrum: Bayesian Estimation of reduced Reparametrized Unified Model
(rRUM) with Gibbs Sampling}
\description{
Implementation of Gibbs sampling algorithm for Bayesian Estimation
of reduced Reparametrized Unifed Model (rRUM), described by Culpepper and
Hudson (2017) <doi: 10.1177/0146621617707511>.
}
\details{
Implemention of the reduced Reparametrized Unified Model
}
\references{
Culpepper, S. A. & Hudson, A. (In Press). An improved strategy for Bayesian 
estimation of the reduced reparameterized unified model. Applied
Psychological Measurement.

Hudson, A., Culpepper, S. A., & Douglas, J. (2016, July). Bayesian estimation
of the generalized NIDA model with Gibbs sampling. Paper presented at the
annual International Meeting of the Psychometric Society, Asheville, North
Carolina.
}
\author{
\strong{Maintainer}: Steven Culpepper \email{sculpepp@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item Aaron Hudson \email{awhudson@uw.edu} [copyright holder]
}

}
