% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrum-package.R
\docType{package}
\name{rrum-package}
\alias{rrum-package}
\alias{_PACKAGE}
\title{rrum: Bayesian Estimation of 'RRUM' Model with Gibbs Sampling}
\description{
Implementation of Gibbs sampling algorithm for Bayesian Estimation
    of the Reduced Reparameterized Unified Model ('RRUM'), described by 
    Culpepper and Hudson (2017) <doi: 10.1177/0146621617707511>.
}
\details{
Implemention of a Bayesian estimation for the
reduced Reparameterized Unified Model (rRUM).
}
\references{
Culpepper, S. A. & Hudson, A. (In Press). An improved strategy for Bayesian
estimation of the reduced reparameterized unified model. Applied
Psychological Measurement.

Hudson, A., Culpepper, S. A., & Douglas, J. (2016, July). Bayesian estimation
of the generalized NIDA model with Gibbs sampling. Paper presented at the
annual International Meeting of the Psychometric Society, Asheville, North
Carolina.
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (0000-0003-4226-6176) [copyright holder]
  \item Aaron Hudson \email{awhudson@uw.edu} (0000-0002-9731-2224) [copyright holder]
}

}
